+++ docs/utils.rst
@@ -1,0 +1,11 @@
+Utils
+=====
+
+
+Module Documentation
+--------------------
+.. automodule:: doctor.utils
+    :members:
+    :private-members:
+    :show-inheritance:
+
